requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pika>=1.3.1
python-dotenv>=1.0.0
